Prevent bypassing locked ships restriction on MoveChar simply by using different casing
Movechar is case-insensitive but the lock checking was not (until now).
This patch also takes the opportunity to change a couple of useless loops
to direct lookups, but the ToLower changes are the important ones.

Antonio reported this problem.

Also fix the case where a character is inserted into the locked ships
database with incorrect casing, causing basic lock functionality (prevent
selling of certain things) to not be applied.