Help Expanded (#25)
New Command: /helpmenu

This command allows user to bring up message boxes that contain helpful information. Aside from two, these are all controlled by a cfg file. Typing the command without any params will cause it to list all possible combinations. The two that are hardcoded in are "/helpmenu rules" & "/helpmenu start". This is done because there are existing infocards in game for these, and doing it this way will circumvent the HkChangeIDSString bug that causes xml strings to break.