Dockingmodules rework (#32)
O frabjous day! Callooh! Callay! Docking modules are back!

An lot of work went into this by Remnant and now that we actually know where carriers are going to be situated going forward in the game and the meta we can merge this in and turn docking modules back on. Thanks to everyone who put the time into look into this and everyone in the rest of the Discovery team who worked on the carrier reworks.