base: Fix repairs not working properly on stations with certain configurations
Bases with dont_eat==1 && (logic==1 || invulnerable==0) (of which there are 7
in Discovery) would calculate new base_health values when repairing, and
commodities would be consumed, but this would not be set in the game itself.
Then, when CoreModule::Timer next got run, the base_health stored in our plugin
would be discarded and replaced with the value from the game, which had not
been updated at last repairs.
This lead to those 7 bases wasting repair commodities and getting no repairs.