Docking modules plugin v0.4.3 (#90)
This PR supposed to be bug fix only. Main changes:

Forced system switch removed completely. I think that is cause of most nasty bug.
SaveFiles.cpp removed, this part didn't its job somehow properly anyway.

This is an attempt to solve the long-standing infinite-jump-bug present relating to this plugin (the exact nature of the bug is not fully known but it is believed to be around system switching). If things go well we can leave it running on the server, make improvements to it, and move it out of the broken plugins folder in VS.