Ship Class Item restrictions: Check for more than one restricted item
The old code was looking at the first restricted item, and if the ship was
whitelisted, it stopped there and did not check for any other restricted items.
This was missing gunboats with forward guns (restricted, but whitelisted for
the class) also mounting battleship scanners (restricted, and not permitted on
gunboats).