Fix issues around special jumps not triggering JumpInComplete
Karlotta found that if you used a jump drive to jump into a system with a POB,
your normal reputation to that POB would not apply. Instead it would always be
neutral.
In addition, I noticed that SensorSystem stuff did not activated when using a
jump drive.

This was because our hack-around-FL-memory trick to change player's destination
system when jumping doesn't lead to us calling JumpInComplete. Instead,
introduce a plugin communication message that provides the requisite
information.