playercntl: raise ReqRemoveItem priority
To ensure it always runs before the base plugin's callback for this hook - that has priority 15.
If we don't do this, playercntl blocking an item sale to a POB will result in the player being
kicked as one would expect - however the item will still be sold, leading to item duplication.