alley dockrestrict: Don't try to PrintLocalUserCmdText where the player is no longer on the server
We found the most common exception attributed to a plugin was
`Exception in plugin 'alley' in HkTimerCheckKick`
The other code run in that hook looked safe but we found this. It appears that
when a player has nodock applied to them and logs off before it expires,
FLHook.log gets spammed with these exceptions.

Kaz wrote this and I approved.