feat(baseplugin): repairing with money
when base is damaged or in siege, it is repaired with repair materials as usual
when it has 100% health, it is repaired with money
FOW is just removed.
any repair mod can be turned on/off now by base owner.
it works precisely only when base is having 100% health now, and just blocks tear/wear damage
with substracting necessary money instead of real repairs.
This change is asked from me by Durandal if I am not mistaken, in January of 2020 year.