Fix various RIGHT_SUPERADMIN checks to not pass people with any less-than-superadmin permissions
RIGHT_SUPERADMIN is 0xFFFFFFFF. So you take someone's rights, bitwise AND it
against that, and what you'll get back is their original rights - which if they
have any, will always be > 0. So they pass the check. This style of check would
work for any other RIGHT_* constant, just not RIGHT_SUPERADMIN